+++ packages/core/src/entries/writers.ts
@@ -2,6 +2,7 @@ export type {
   TraceWriter,
   TraceWriterStats,
   MemoryTraceWriter,
+  FileTraceWriterOptions,
 } from "../writers/index.js";
 
-export { memoryWriter, nullWriter } from "../writers/index.js";
+export { fileWriter, memoryWriter, nullWriter } from "../writers/index.js";